Automate repo updates with AWS Lambda
`app.py` contains the AWS CDK to create a Python Lambda function with
the appropriate permission to log and pull all the IAM Policies.

The `dulwich` module is used for all git operations.

Before running `cdk deploy` an Secret needs to be created in AWS
Managed Secrets and a few environment variables need to be created.
See `app.py` for instructions.

Lambda Function is invoked every hour. To test, deploy and trigger
the Lambda with a test event. Execution can take around a minute. More
memory will speed it up.